# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend modular API for GNyx. (Powered by Kubex Ecosystem)",
        "title": "GNyx API",
        "termsOfService": "https://gnyx.kubex.world/terms",
        "contact": {
            "name": "GNyx API Support",
            "url": "https://gnyx.kubex.world",
            "email": "contact@kubex.world"
        },
        "version": "1.0.0"
    },
    "host": "localhost:5001",
    "basePath": "/",
    "paths": {
...
        "/api/v1/bi/boards/export": {
            "post": {
                "description": "Exports the generated board payloads as a ZIP archive for portability and demo usage.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "Runtime BI"
                ],
                "summary": "Export a generated BI board bundle",
                "parameters": [
                    {
                        "description": "Board export request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.boardExportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ZIP archive",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.authErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/routes.authErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/bi/boards/generate": {
            "post": {
                "description": "Generates a grounded BI board plan and dashboard schema for the supported domain slice.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Runtime BI"
                ],
                "summary": "Generate a BI board",
                "parameters": [
                    {
                        "description": "Board generation request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.boardGenerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.boardGenerateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.authErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/routes.authErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/routes.unifiedErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/bi/catalog/status": {
            "get": {
                "description": "Returns the status of the grounded Sankhya metadata catalog used by BI generation.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Runtime BI"
                ],
                "summary": "Get BI catalog status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bi.CatalogStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/routes.authErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bi.BoardPlan": {
            "type": "object",
            "properties": {
                "assumptions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "board_description": {
                    "type": "string"
                },
                "board_title": {
                    "type": "string"
                },
                "business_goal": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "filters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.BoardPlanFilter"
                    }
                },
                "grounding": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.BoardPlanGroundingRef"
                    }
                },
                "joins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.GroundingJoin"
                    }
                },
                "time_reference": {
                    "$ref": "#/definitions/bi.BoardPlanTimeReference"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "widgets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.BoardPlanWidget"
                    }
                }
            }
        },
        "bi.BoardPlanDataSource": {
            "type": "object",
            "properties": {
                "expected_granularity": {
                    "type": "string"
                },
                "main_table": {
                    "type": "string"
                },
                "sql": {
                    "type": "string"
                }
            }
        },
        "bi.BoardPlanFilter": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "string"
                },
                "filter_group": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "bi.BoardPlanGroundingRef": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "relevant_fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "row_count": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "bi.BoardPlanTimeReference": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "bi.BoardPlanWidget": {
            "type": "object",
            "properties": {
                "assumptions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "data_source": {
                    "$ref": "#/definitions/bi.BoardPlanDataSource"
                },
                "display": {
                    "$ref": "#/definitions/bi.BoardPlanWidgetDisplay"
                },
                "grounding": {
                    "$ref": "#/definitions/bi.BoardPlanWidgetGrounding"
                },
                "id": {
                    "type": "string"
                },
                "metric_goal": {
                    "type": "string"
                },
                "size": {
                    "$ref": "#/definitions/bi.BoardPlanWidgetSize"
                },
                "subtitle": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/bi.WidgetType"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "bi.CatalogDatasetStatus": {
            "type": "object",
            "properties": {
                "dataset_name": {
                    "type": "string"
                },
                "last_loaded_at": {
                    "type": "string"
                },
                "row_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "table_name": {
                    "type": "string"
                }
            }
        },
        "bi.CatalogStatus": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "datasets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.CatalogDatasetStatus"
                    }
                },
                "domain": {
                    "type": "string"
                },
                "empty_datasets": {
                    "type": "integer"
                },
                "failed_datasets": {
                    "type": "integer"
                },
                "last_loaded_at": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "ready_datasets": {
                    "type": "integer"
                },
                "schema_name": {
                    "type": "string"
                },
                "source_system": {
                    "type": "string"
                },
                "sync_command_example": {
                    "type": "string"
                },
                "total_datasets": {
                    "type": "integer"
                }
            }
        },
...
        "bi.DashboardSchema": {
            "type": "object",
            "properties": {
                "dashboardId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "widgets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.DashboardWidget"
                    }
                }
            }
        },
        "bi.DashboardWidget": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/bi.DashboardWidgetConfig"
                },
                "id": {
                    "type": "string"
                },
                "size": {
                    "$ref": "#/definitions/bi.DashboardWidgetSize"
                },
                "type": {
                    "$ref": "#/definitions/bi.WidgetType"
                }
            }
        },
...
        "bi.DashboardWidgetSize": {
            "type": "object",
            "properties": {
                "cols": {
                    "type": "integer"
                },
                "rows": {
                    "type": "integer"
                }
            }
        },
        "bi.GroundingContext": {
            "type": "object",
            "properties": {
                "catalog_schema": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "domain_options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.GroundingOption"
                    }
                },
                "filter_candidates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.GroundingFilter"
                    }
                },
                "joins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.GroundingJoin"
                    }
                },
                "primary_tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.GroundingTable"
                    }
                },
                "recommended_widgets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.WidgetType"
                    }
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "bi.GroundingField": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "option_count": {
                    "type": "integer"
                }
            }
        },
        "bi.GroundingFilter": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "filter_group": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "bi.GroundingJoin": {
            "type": "object",
            "properties": {
                "basis": {
                    "type": "string"
                },
                "join_type": {
                    "type": "string"
                },
                "left_column": {
                    "type": "string"
                },
                "left_table": {
                    "type": "string"
                },
                "right_column": {
                    "type": "string"
                },
                "right_table": {
                    "type": "string"
                }
            }
        },
        "bi.GroundingOption": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "label": {
                    "type": "string"
                },
                "option_label": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "bi.GroundingTable": {
            "type": "object",
            "properties": {
                "date_columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "physical_schema": {
                    "type": "string"
                },
                "relevant_fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bi.GroundingField"
                    }
                },
                "row_count": {
                    "type": "integer"
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "bi.WidgetType": {
            "type": "string",
            "enum": [
                "kpi",
                "chart_bar",
                "chart_donut",
                "data_table"
            ],
            "x-enum-varnames": [
                "WidgetTypeKPI",
                "WidgetTypeChartBar",
                "WidgetTypeChartDonut",
                "WidgetTypeDataTable"
            ]
        },
...
        "routes.authErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "routes.boardExportRequest": {
            "type": "object",
            "properties": {
                "dashboard_schema": {
                    "$ref": "#/definitions/bi.DashboardSchema"
                },
                "domain": {
                    "type": "string"
                },
                "fallback_reason": {
                    "type": "string"
                },
                "generation_mode": {
                    "type": "string"
                },
                "grounding_context": {
                    "$ref": "#/definitions/bi.GroundingContext"
                },
                "model": {
                    "type": "string"
                },
                "plan": {
                    "$ref": "#/definitions/bi.BoardPlan"
                },
                "prompt": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "usage": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
        "routes.boardGenerateRequest": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "max_widgets": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
        "routes.boardGenerateResponse": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/routes.providerAttempt"
                    }
                },
                "dashboard_schema": {
                    "$ref": "#/definitions/bi.DashboardSchema"
                },
                "domain": {
                    "type": "string"
                },
                "fallback_reason": {
                    "type": "string"
                },
                "generation_mode": {
                    "type": "string"
                },
                "grounding_context": {
                    "$ref": "#/definitions/bi.GroundingContext"
                },
                "model": {
                    "type": "string"
                },
                "plan": {
                    "$ref": "#/definitions/bi.BoardPlan"
                },
                "provider": {
                    "type": "string"
                },
                "raw_provider_json": {
                    "type": "string"
                },
                "usage": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
...
        "routes.providerAttempt": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
...
        "routes.unifiedErrorResponse": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/routes.providerAttempt"
                    }
                },
                "error": {
                    "type": "string"
                },
                "error_class": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Provide \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}